LP#1579219: set up mod_perl as an Evergreen dependency
This patch, in conjunction with another one for OpenSRF,
set mod_perl up as a dependency installed when Evergreen is installed,
not when OpenSRF is installed, as OpenSRF itself does not require
mod_perl.

Minor tweak to Galen's original patch to include mod_perl in the list
for apache mods to be activated by default.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <ben@evergreener.net>