LP#1613381: combine two tables in patron notification preferences
This patch consolidates the two tables in the public catalog
notification preferences page into one table, as patrons have
no reason to care about the distinction between A/T opt-in
settings and the hold, phone, and SMS user settings.

This patch also adds a <label> element for opt-in setting
names.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>