LP#1494750 - Extra closing curly bracket in style.css disables following css
Which is the .sort_deemphasize class, so the modification of the non-filing
characters never happens in the sort results.

Test Plan:

Needed: Patron account with checkout history that contains titles with
non filing characters.

Before change, view the checkout history and sort by title.  Note that the
leading non filing characters look the same as the rest of the title.

After change, view the checkout history and sort by title.  Note that the
leading non filing characters now have a different style than the rest
of the title.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Jeanette Lundgren <jlundgren@cwmars.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>