LP#1614807: Fix spacing issues in responsive design for My Account screens
This commit fixes the following spacing issues that are found in the My Account
screens when loaded on a small-screen device:

* When a field has no value, the text in the table header, which displays along
 the left, is cut off at the bottom. The Notes field on the Holds page is one
place where this issue is commonly seen. With the fix, the headers are slightly
out of alignment with the field data, but this is still an improvement over
cutting off text on the screen.

* If the only title(s) listed on the screen are short, requiring little width
in a table cell, the text in the table header runs into field data. A good
example in the Concerto data is if the user only has a check out for record ID
2 (Le concerto) on their account.

* The checkboxes in the Check Out History and Items on Hold screen were pushed
to the right. We fix it by giving the cells with a checkbox a "checkbox_column"
class, which was already in use by the checkbox on the Items Out screen.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>