LP#1467663 webstaff: dedicated workstation admin page
Adds a new standalone workstations admin page, accessible from the main
Workstation Administration page.

Adds the ability to delete a workstation by name both within the page
and via URL (with permission).

Fixes the "Use Now" button so that it actually logs the user out before
redirecting to the login page with the desired workstation
pre-selected.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>