LP#1467663 webstaff: login requires valid workstation
Hide the workstation selector when no workstations are registered.
After successful login, direct the user to the new workstation admin
page to create a new workstation.

After successful login with an invalid workstation, direct the user
to the workstation admin page, issuing a 'remove' command to un-register
the offending WS.  On the WS admin page, the user can create a new
workstation or select from their existing workstations.

Any attempt to access a browser client interface (minus the WS admin
page) without a valid workstation will cause the page to redirect to
the workstation admin page.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>