webstaff: implement patron merge interface
This patch adds a 'Merge Patrons' button to the patron
search grid. If the user selects two patron records, the
button can be clicked to present a dialog that allows
the user to pick a lead record and confirm a merge of the
patrons.

This patch also adds an egPatronSummary directive that
uses the existing patron summary template with a couple
modifications.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>