implement item attributes editor AKA volume/copy template editor
Add a specialization of the volume/copy editor that is restricted
just to managing volume/copy templates; this can be accessed
at the path /eg/staff/cat/volcopy/edit_templates.  A link is added
to the local administration splash page with the label "Volume/Copy
Template Editor"

In attempt to improve terminology, on the local administration page,
the previous "Copy Template Editor" is now called "Serial Copy Template
Editor".

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>