Fix purge_user_activity.pg live test
The purge_user_activity.pg test inserts rows into actor.usr_activity
in a random date order, but the changes in bug 1570909 created a trigger
that overwrites the previous entry for a transient type, causing the
test to fail.  We reorder the usr_activity rows so that the most recent
row will be in place.

Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Jason Stephenson <jason@sigio.com>