webstaff: tweaks to egPrint
This patch attempts to unbreak receipt and CSV printing
by fetching and injecting the print CSS into a style element,
as a link element in the print div doesn't seem to get
processed.  It also automatically clears the content
of the print div so that after you print a receipt, the
browser print command will print the page, not the receipt.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>