LP 1473054: No-op Checkin Should Not Always Float a Floating Copy.
Prior to this patch, a no-op checkin would not check if a copy should
float using the evergreen.can_float() function.  It would simply float
the copy to the org_unit where the checkin happened if the copy had a
float value set.

This commit changes the behavior to use evergreen.can_float() to
determine if the copy should float to the org_unit where the checkin
happened.  This change makes the behavior consistent with a regular
checkin.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>