LP#1117808: new Pg function to calculate MARC record merges
This patch adds a new stored function, vandelay.merge_record_xml_using_profile(),
that when passed two MARCXML strings and the ID of a merge profile.
returns the MARCXML string representing the results of the record merge.

This is meant for situations where we want to know the results
of a merge, but not immediately overlay a record in the database.

pgTAP tests are available in t/marc_merge.pg.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>