LP#1663374: Allow metarecord searching by default
Before the TPAC, a site could configure the OPAC to use metarecord searching
by default, via a configuration file.  Here we bring that back.

A new setting is present in templates/opac/parts/config.tt2 to enable this
feature.  By setting this to a true value (normally 1) the TPAC will silently
include the #metabib search modifier in the search form on any interfaces that
do not have a UI component that allows the user to control the setting.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>