LP 1648918: Remove Workstation at Root of User's Permission Tree
If you register a workstation at the root of the user's permission
tree in the webstaff client, you are not able to later remove it.  The
Remove button is dimmed.  You can remove a workstation registered at
any other org_unit where the user has the REGISTER_WORKSTATION
permission.

To reproduce this bug, login as the admin user and register a
workstation at the CONS org_unit.  Then, go to Workstation
Administration and see that the Remove button is dimmed.

To test the patch, do the same, but see that the Remove button is not
dimmed.  When you click it, the workstation is also removed.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Michele Morgan <mmorgan@noblenet.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>