LP#1650807: fix rollover_phone_to_print.pl to ack failed calls
This patch fixes a bug in rollover_phone_to_print.pl that kept
failed callfiles from being moved.

The list of new event ID's that the script created was being sent to the
mediator, which couldn't do anything with them.  The original event ID's
need to be sent to have those call files moved.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>