LP#1117808: teach record bucket merge about merge profiles
This patch adds a widget to the record bucket merge dialog
to allow the user to select a MARC merge profile to use
during the merge.

After the user has chosen a lead record, the result of the merge
is displayed in the left-hand pane, and is updated whenever the user

- changes the selected merge profile
- swaps in a different lead record
- removes a subordinate record from consideration

As before, the user can choose to edit the lead record, but note that
the version that is edited is the /original/ version of the
lead record, with any changes due to the merge profile
being applied after the edit.

Since there can be more than one subordinate record in play, the
result of the merge is calculated by merging the first subordinate
record into the lead record, then in the second subordinate record,
and so forth.

Merge profiles that have a 'preserve' specification are excluded
from selection, as such profiles have the effect of swapping what
is considered the lead record for the purpose of the MARC merge.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>