LP#1117808: teach Z39.50 import about merge profiles
This patch adds a widget to the Z39.50 overlay dialog to
allow the user to select a MARC merge profile to use during
the merge. For the purpose of applying the merge profile,
the incoming record from the Z39.50 server is treated as
the target of the merge, as the user is assumed to be picking
a Z39.50 record that is "better" overall but may need to be
supplement with fields from the incumbent record.

As before, the user can choose to edit the incoming record, but
note that the version that is edited is the /original/ version of
the Z39.50 record, with any changes due to the merge profile being
applied after the edit.

Merge profiles that have a 'preserve' specification are excluded
from selection, as such profiles have the effect of swapping what
is considered the lead record for the purpose of the MARC merge

This patch also tweaks the dialog to use the egRecordBreaker directive
to display the incumbent record, as that permits easier copy-and-pasting
if the user chooses to edit the Z39.50 record.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>