webstaff: booking options from Patron interface
Includes some workarounds for xulG errors, and lets the booking interfaces
receive the patron barcode via a query param instead of xulG.  The XUL client
spawns a new tab for these interfaces, but we stick with web norms and load in
the current tab (control+click will open a new tab).  This does, however, blow
away the patron interface; it doesn't embed like most of the options from the
Other menu.  One other change: we don't have support for nested drop-down menus
in the web client, so I've flattened the entries like so:

* Booking: Create or Cancel Reservations
* Booking: Pick Up Reservations
* Booking: Return Reservations

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>