Docs: 2.11 Release Note corrections and clarifications.
Fixes several typos in the 2.11 Release Notes and in the template for release
notes acknowledgements. Also resets the acknowledgements as TODOs for the
next release. Added clarification to the email checkout receipts to let users
know that email receipts are not available in the XUL client, but only in the
web client.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>