LP#1646166 À la carte Hatch, on-call settings, strict access.
1. Hatch now supports a al carte features instead of requiring all-or-none
functionality.  Supported features currently include printing, settings,
and offline.  (Note: offline handling pending merge of offline UI code).

2. Adds support for on-call setting keys.  On-Call keys are those that
can be set/get/remove'd from localStorage when Hatch is not avaialable,
even though Hatch is configured as the primary storage location for the
key in question.

The initital target use case for on-call keys are those that allow the
user to login and perform basic admin tasks (like disabling Hatch) even
when Hatch is down.  AKA Browser Staff Run Level 3.

3. egHatch no longer attempts requests at Hatch, falling through to
local requests when Hatch fails.  With the exception of on-call keys
(above), either Hatch is used or local requests are used, depending on
the configuration.  The goal is to prevent any unintended and
confusing blending of local and remote data.  In other words, if Hatch
is broken, it needs to be fixed or disabled for regular work flow to
continue.

4. Hatch now has a dedicated UI under workstation administration.

5. Workstation admin splash page rearranged to take advantage of more
horizontal space and avoid pushing so many options down the page.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>