LP#1657885: Account for new bundling/chunking logic in OpenSRF 2.5+
When chunking was renamed bundling and actually chunking added in OpenSRF 2.5,
the few places in Evergreen that tried to make use of the old mechanism
directly now break. The most obvious breakage is in the alternate printable
hold pull list, which we fix here.  Evidence of other broken code should
be fixed as needed, though spots to look our for are z39.50 results and
Vandelay processing.

To test:

 1) In Evergreen 2.11 running on top of OpenSRF 2.5+, attempt to use the
    alt pull list printing interface at a location with many holds on their
    pull list.  The progress bar will spin forever.
 2) After applying this patch, do the same.  The interface should work
    quickly.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>