LP#1312837 - Item Status - Alternate View - Holds/Transit tab: Transit and Hold  information does not refresh
When switching between copies that have different transit and hold information,
the data in the hold/transit tab doesn't get cleared when there isn't a new hold
or transit to clear it.

Test Plan:

Acquire the barcodes for 3 copies, one that is available, on that is in transit
and one that is captured for a hold.

Before Change:
  1. Open item status alternate view and switch to hold/copy tab.  Note that the
     labels default to saying that the item is in transit and captured for a
     hold before any barcodes are scanned.
  2. Enter the available item and note that the labels changed to not in transit
     , not captured.
  3. Scan in the In Transit for Captured item and note that the labels don't
     change when they should.
  4. Scan in the available item and note that the hold or transit data isn't
     cleared.

After Change:
  1. Enter the barcodes for the 3 copies with different statuses and note that
     the Holds/Transit information now updates correctly.

Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>