LP#1494748: Decrease the input size on Firefox by a smidge
The larger scale works for the input works well in Chrome, but came across as
too large in blurry in Firefox. Decreasing the Firefox size by a smidge so
that they still appear larger but don't become blurry. Firefox only seemed to
recognize the last transform, regardless of prefix, so I moved the one intended
for Firefox lower in the CSS list.

Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Josh Stompro <stomproj@larl.org>