LP1629108 metarecord_constituent_result_reroute
This patch will route the metasearch logic through the "standard" search logic
in order to leverage the heavy use of filters and other features.

A column is introduced to unapi.mmr_mr to include the constituent bibs in the
return. A tweak was required in the template toolkit code to take advantage of the
new payload. This enables TT to decide which icons should be displayed when search
results are filtered.

Signed-off-by: blake <blake@mobiusconsortium.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>