LP#1662541: several new subject browse index definitions
This patch adds several config.metabib_field definitions
that are specifically for subject browses:

subject / topic_browse
subject / geographic_browse
subject / temporal_browse

These new definitions index the entire heading as a unit,
so a heading like this:

=650 \0$aCats$zUnited States$vCorrespondence.

will end up in the subject/topic_browse index as a single
entry for

Cats -- United States -- Correspondence

rather than separate browse entries for "Cats" as a subject/topic,
and "United States" as a subject/geographic.

Since the new *_browse indexes are meant to replace the browse
functions of the existing subject/topic, subject/geographic, and
subject/temporal indexes, the browse flag is turned off for the
older definitions.

This patch intentionally does not include an entry for subject/name_browse,
as additional coding would be required to get the display form of the
heading to be punctuated correctly.

To test
-------
[1] After applying the patch, perform subject browses in the
    OPAC and staff interfaces and verify that fully composed
    headings from bib records are now displayed as browse entries.

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>